+++ src/function/script_button.ts
@@ -11,17 +11,14 @@ export function _getButtonEvent(this: Window, button_name: string): string {
   return `${String(_getScriptId.call(this))}_${button_name}`;
 }
 
-export function getScriptButtons(script_id: string): ScriptButton[] {
-  if (!script_id) {
-    throw new Error('脚本ID不能为空');
-  }
-  return ScriptManager.getInstance().getScriptButton(script_id);
+export function _getScriptButtons(this: Window): ScriptButton[] {
+  return ScriptManager.getInstance().getScriptButton(_getScriptId.call(this));
 }
 
-export function replaceScriptButtons(script_id: string, buttons: ScriptButton[]): void {
-  if (!script_id) {
-    throw new Error(`脚本ID不能为空`);
-  }
+export function _replaceScriptButtons(this: Window, script_id: string, buttons: ScriptButton[]): void;
+export function _replaceScriptButtons(this: Window, buttons: ScriptButton[]): void;
+export function _replaceScriptButtons(this: Window, param1: string | ScriptButton[], param2?: ScriptButton[]): void {
+  const script_id = _getScriptId.call(this);
 
   const script = ScriptManager.getInstance().getScriptById(script_id);
   if (!script) {
@@ -30,16 +27,24 @@ export function replaceScriptButtons(script_id: string, buttons: ScriptButton[])
 
   const type = ScriptData.getInstance().getScriptType(script);
 
-  script.buttons = buttons;
+  script.buttons = typeof param1 === 'string' ? param2! : param1;
   ScriptManager.getInstance().setScriptButton(script, type);
 }
 
-export function appendInexistentScriptButtons(script_id: string, buttons: ScriptButton[]): void {
-  const script_buttons = getScriptButtons(script_id);
+export function _appendInexistentScriptButtons(this: Window, script_id: string, buttons: ScriptButton[]): void;
+export function _appendInexistentScriptButtons(this: Window, buttons: ScriptButton[]): void;
+export function _appendInexistentScriptButtons(
+  this: Window,
+  param1: string | ScriptButton[],
+  param2?: ScriptButton[],
+): void {
+  const buttons = typeof param1 === 'string' ? param2! : param1;
+
+  const script_buttons = _getScriptButtons.call(this);
   const inexistent_buttons = buttons.filter(button => !script_buttons.some(sb => sb.name === button.name));
   if (inexistent_buttons.length === 0) {
     return;
   }
 
-  replaceScriptButtons(script_id, [...script_buttons, ...inexistent_buttons]);
+  _replaceScriptButtons.call(this, [...script_buttons, ...inexistent_buttons]);
 }